move basic checkers to Test.QuickCheck.Checkers
