+++ checkers-numhask/checkers-numhask.cabal
@@ -12,15 +12,17 @@ category:            Test
 build-type:          Simple
 extra-source-files:  README.md
 cabal-version:       >=1.10
-tested-with:         GHC == 7.10.2
+tested-with:         GHC == 8.4.1
 
 library
   default-language:    Haskell2010
   ghc-options:         -Wall
   hs-source-dirs:      src
-  exposed-modules:     Test.QuickCheck.Checkers.NumHask
+  exposed-modules:     Test.QuickCheck.Checkers
+                       Test.QuickCheck.Checkers.NumHask
                        Test.QuickCheck.Utils
   build-depends:       base >= 4.7 && < 5
+                     , numhask >= 0.2
                      , QuickCheck
